fix: modularize codebase
